backend_python/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
supabase==2.0.3
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0

# Redis Cache
redis==5.0.1
hiredis==2.2.3

# Email & SMS
sendgrid==6.10.0
twilio==8.10.2

# Payment Gateways
stripe==7.5.0

# Image Processing
pillow==10.1.0
python-magic==0.4.27

# HTTP Client
httpx==0.24.1
aiohttp==3.9.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.0.3

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.0